+++ src/domain/locking/entities/schemas/__tests__/campaign.schema.spec.ts
@@ -1,6 +1,5 @@
 import { campaignBuilder } from '@/domain/locking/entities/__tests__/campaign.builder';
-import { CampaignSchema } from '@/domain/locking/entities/schemas/campaign.schema';
-import { faker } from '@faker-js/faker';
+import { CampaignSchema } from '@/domain/locking/entities/campaign.entity';
 import { ZodError } from 'zod';
 
 describe('CampaignSchema', () => {
@@ -12,37 +11,18 @@ describe('CampaignSchema', () => {
     expect(result.success).toBe(true);
   });
 
-  it('should coerce periodStart to a date', () => {
-    const periodStart = faker.date.recent();
-    const campaign = campaignBuilder()
-      .with('periodStart', periodStart.toISOString() as unknown as Date)
-      .build();
+  it.each([
+    'periodStart' as const,
+    'periodEnd' as const,
+    'lastUpdated' as const,
+  ])(`should coerce %s to a date`, (field) => {
+    const campaign = campaignBuilder().build();
 
     const result = CampaignSchema.safeParse(campaign);
 
-    expect(result.success && result.data.periodStart).toBeInstanceOf(Date);
-  });
-
-  it('should coerce periodEnd to a date', () => {
-    const periodEnd = faker.date.recent();
-    const campaign = campaignBuilder()
-      .with('periodEnd', periodEnd.toISOString() as unknown as Date)
-      .build();
-
-    const result = CampaignSchema.safeParse(campaign);
-
-    expect(result.success && result.data.periodEnd).toBeInstanceOf(Date);
-  });
-
-  it('should coerce lastUpdated to a date', () => {
-    const lastUpdated = faker.date.recent();
-    const campaign = campaignBuilder()
-      .with('lastUpdated', lastUpdated.toISOString() as unknown as Date)
-      .build();
-
-    const result = CampaignSchema.safeParse(campaign);
-
-    expect(result.success && result.data.lastUpdated).toBeInstanceOf(Date);
+    expect(result.success && result.data[field]).toStrictEqual(
+      new Date(campaign[field]),
+    );
   });
 
   it('should not validate an invalid campaign', () => {
